requirements.txt
botocore
boto3
blomp-api
fastapi
uvicorn[standard]  # Uvicorn is the high-performance ASGI server
pydantic
motor
gunicorn
redis[async]
orjson
pydantic[email]
pycryptodome
firebase-admin
requests
beautifulsoup4
httpx
psycopg2-binary
python-swiftclient
python-keystoneclient
webdavclient3
selenium
google-genai
PyJWT
passlib[bcrypt]